+++ src/paychangu/resources/__init__.py
@@ -2,16 +2,20 @@
 
 from .bills import BillsResource
 from .cards import CardsResource
+from .connect import ConnectResource
 from .direct_charge import DirectChargeResource
 from .payments import PaymentsResource
 from .payouts import PayoutsResource
+from .virtual_accounts import VirtualAccountsResource
 from .wallet import WalletResource
 
 __all__ = [
     "BillsResource",
     "CardsResource",
+    "ConnectResource",
     "DirectChargeResource",
     "PaymentsResource",
     "PayoutsResource",
+    "VirtualAccountsResource",
     "WalletResource",
 ]